+++ homework_6/src/main/java/ru/otus/raukhvarger/homework_6/jpa/repository/GenreRepository.java
@@ -5,7 +5,7 @@ import ru.otus.raukhvarger.homework_6.jpa.entity.GenreEntity;
 import java.util.List;
 
 public interface GenreRepository {
-    GenreEntity getById(Integer genreId);
+    GenreEntity getById(Long genreId);
 
     GenreEntity getByName(String genreName);
 
@@ -13,7 +13,7 @@ public interface GenreRepository {
 
     void update(GenreEntity genreEntity);
 
-    void deleteById(Integer genreId);
+    void deleteById(Long genreId);
 
     List<GenreEntity> getAll();
 }
